update for multi-device
