+++ ppfleetx/utils/device.py
@@ -1,0 +1,39 @@
+import paddle
+
+def get_device_and_mapping():
+    """
+        Return gpu type and name-bool mapping implifying which type is supported.
+    """
+    suppoted_device_map = {
+        "gpu": paddle.is_compiled_with_cuda(),
+        "xpu": paddle.is_compiled_with_xpu(),
+        "rocm": paddle.is_compiled_with_rocm(),
+        "npu": paddle.is_compiled_with_npu(),
+        "cpu": True
+    }
+    for d, v in suppoted_device_map.items():
+        if v:
+            return d, suppoted_device_map
+
+
+def get_device():
+    """
+        Return the GPU Type, including 'gpu'(for rocm and gpu), 'npu', 'xpu', 'cpu'.
+    """
+    d, _ = get_device_and_mapping()
+    return d
+
+
+def synchronize():
+    """
+    synchronize device, return True if suceed synchronized, Otherwise rtunr False
+    """
+    if paddle.is_compiled_with_cuda():
+        paddle.device.cuda.synchronize()
+        return True
+    elif paddle.is_compiled_with_xpu():
+        paddle.device.xpu.synchronize()
+        return True
+    else:
+        logger.warning("The synchronizzation is only supported on cuda and xpu now.")
+    return False
