+++ ppfleetx/utils/log.py
@@ -27,7 +27,7 @@ import colorlog
 from colorama import Fore
 
 import paddle
-from .version import get_device
+from .device import get_device, synchronize
 
 loggers = {}
 
@@ -180,13 +180,10 @@ def advertise():
 
 
 def get_timestamp():
-    if paddle.is_compiled_with_cuda():
-        paddle.device.cuda.synchronize()
-    elif paddle.is_compiled_with_xpu():
-        paddle.device.xpu.synchronize()
+    if synchronize():
+        return time.time()
     else:
-        logger.warning(f"the time stamp is only supported on cuda and xpu. "
-        f"However current supported is {get_device()}, which may result uncorrect time")
+        logger.warning(f"Device synchronizing failed, which may result uncorrect time")
     return time.time()
 
 def convert_timestamp_to_data(timeStamp):
